+++ lib/src/fields/text_field.dart
@@ -1,6 +1,6 @@
 import 'dart:ui' as ui show BoxHeightStyle, BoxWidthStyle;
 
-import 'package:flutter/foundation.dart' show defaultTargetPlatform, kIsWeb;
+import 'package:flutter/foundation.dart' show defaultTargetPlatform;
 import 'package:flutter/gestures.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
@@ -96,7 +96,7 @@ class _TextFieldSelectionGestureDetectorBuilder
     extends TextSelectionGestureDetectorBuilder {
   _TextFieldSelectionGestureDetectorBuilder({
     required _TextFieldState state,
-  })   : _state = state,
+  })  : _state = state,
         super(delegate: state);
 
   final _TextFieldState _state;
@@ -1145,7 +1145,7 @@ class _TextFieldState extends State<TextField>
                     CupertinoIcons.clear_thick_circled,
                     size: 18.0,
                     color:
-                        DynamicColorX.macosResolve(_kClearButtonColor, context),
+                        MacosDynamicColor.resolve(_kClearButtonColor, context),
                   ),
                 ),
               ),
@@ -1195,21 +1195,21 @@ class _TextFieldState extends State<TextField>
     final MacosThemeData themeData = MacosTheme.of(context);
 
     final TextStyle? resolvedStyle = widget.style?.copyWith(
-      color: DynamicColorX.maybeMacosResolve(widget.style?.color, context),
-      backgroundColor: DynamicColorX.maybeMacosResolve(
+      color: MacosDynamicColor.maybeResolve(widget.style?.color, context),
+      backgroundColor: MacosDynamicColor.maybeResolve(
         widget.style?.backgroundColor,
         context,
       ),
     );
 
-    final textStyle = themeData.typography!.body!.merge(resolvedStyle);
+    final textStyle = themeData.typography.body.merge(resolvedStyle);
 
     final resolvedPlaceholderStyle = widget.placeholderStyle?.copyWith(
-      color: DynamicColorX.maybeMacosResolve(
+      color: MacosDynamicColor.maybeResolve(
         widget.placeholderStyle?.color,
         context,
       ),
-      backgroundColor: DynamicColorX.maybeMacosResolve(
+      backgroundColor: MacosDynamicColor.maybeResolve(
         widget.placeholderStyle?.backgroundColor,
         context,
       ),
@@ -1220,13 +1220,13 @@ class _TextFieldState extends State<TextField>
     final Brightness keyboardAppearance =
         widget.keyboardAppearance ?? MacosTheme.brightnessOf(context);
     final Color cursorColor =
-        DynamicColorX.maybeMacosResolve(widget.cursorColor, context) ??
-            themeData.primaryColor!;
+        MacosDynamicColor.maybeResolve(widget.cursorColor, context) ??
+            themeData.primaryColor;
     final Color disabledColor =
-        DynamicColorX.macosResolve(_kDisabledBackground, context);
+        MacosDynamicColor.resolve(_kDisabledBackground, context);
 
     final Color? decorationColor =
-        DynamicColorX.maybeMacosResolve(widget.decoration?.color, context);
+        MacosDynamicColor.maybeResolve(widget.decoration?.color, context);
 
     final BoxBorder? border = widget.decoration?.border;
     Border? resolvedBorder = border as Border?;
@@ -1235,7 +1235,7 @@ class _TextFieldState extends State<TextField>
         return side == BorderSide.none
             ? side
             : side.copyWith(
-                color: DynamicColorX.macosResolve(side.color, context),
+                color: MacosDynamicColor.resolve(side.color, context),
               );
       }
 
@@ -1255,10 +1255,7 @@ class _TextFieldState extends State<TextField>
     );
 
     final BoxDecoration? focusedDecoration = widget.focusedDecoration?.copyWith(
-      border: Border.all(
-        width: 3.0,
-        color: themeData.primaryColor ?? CupertinoColors.systemBlue,
-      ),
+      border: Border.all(width: 3.0, color: themeData.primaryColor),
     );
 
     final focusedPlaceholderDecoration = focusedDecoration?.copyWith(
@@ -1285,7 +1282,7 @@ class _TextFieldState extends State<TextField>
     );
 
     final Color selectionColor =
-        MacosTheme.of(context).primaryColor!.withOpacity(0.2);
+        MacosTheme.of(context).primaryColor.withOpacity(0.2);
 
     final Widget paddedEditable = Padding(
       padding: widget.padding,
@@ -1333,7 +1330,7 @@ class _TextFieldState extends State<TextField>
             cursorOffset: cursorOffset,
             paintCursorAboveText: true,
             autocorrectionTextRectColor: selectionColor,
-            backgroundCursorColor: DynamicColorX.macosResolve(
+            backgroundCursorColor: MacosDynamicColor.resolve(
               CupertinoColors.inactiveGray,
               context,
             ),
@@ -1392,17 +1389,9 @@ class _TextFieldState extends State<TextField>
       ),
     );
 
-    child = MouseRegion(
+    return MouseRegion(
       cursor: SystemMouseCursors.text,
       child: child,
     );
-
-    if (kIsWeb) {
-      return Shortcuts(
-        shortcuts: scrollShortcutOverrides,
-        child: child,
-      );
-    }
-    return child;
   }
 }